Fix data divergence in Arrangement -> Apply Overlay workflow
- Coordinate Precision: Use float for overlay coordinates in ArrangementConfig.
- Sampling Precision: Use PIL's 'box' parameter in resize() for sub-pixel accuracy.
- Dynamic Tile Sizes: Update ArrangementBridge to use tile dimensions from processor.
- Indexing Consistency: Pass explicit tiles_per_row to _update_tile_data_from_modified_tiles.
- Workflow Alignment: Standardize on canvas-coordinate sampling for overlays.

+++ core/arrangement_persistence.py
@@ -40,8 +40,8 @@ class ArrangementConfig:
     grid_mapping: dict[str, dict[str, str]] = field(default_factory=dict)
     # Overlay state (added in v1.1)
     overlay_path: str | None = None
-    overlay_x: int = 0
-    overlay_y: int = 0
+    overlay_x: float = 0.0
+    overlay_y: float = 0.0
     overlay_opacity: float = 0.5
     overlay_visible: bool = True
     # Timestamps
